Move `Card` component to `card` subfolder
