+++ packages/components/src/card/card/index.js
@@ -1,0 +1,2 @@
+export { default } from './component';
+export { useCard } from './hook';
