+++ packages/components/src/card/index.js
@@ -1,6 +1,5 @@
-export { default as Card } from './component';
+export { default as Card, useCard } from './card';
+
 export { default as CardBody } from './body';
 export { default as CardFooter } from './footer';
 export { default as CardHeader } from './header';
-
-export * from './hook';
